Removed Trie and used a vector for the routes.
Even though the Trie increased the efficiency, it made it harder (or impossible) to match some routes, for example /a/*/b with /a/x/y/z/b, since the Trie matched part by part.
I think it is still doable and maybe I will come back on this, but for now I want to keep it easy.