Create custom Network stream to allow flexible retrieval of characters from the network.
The previous implementation had the critical flow of interpreting TCP packets as HTTP requests; however, TCP is a stream-based protocol and packets have therefore no meaning. To fix the problem, and be able to consistently retrieve the entire packet before parsing and handling, a network stream was created; this way, the previous code using std::istream can be left untouch, and the stream will automatically retrieve data from the network when needed. Clearly, this assumes that the client is well-behaved i.e. provides as much data as it is supposed to do.