Implement keep-alive mechanism.
Instead of spawning a thread for each request, create a fixed number of worker threads and let them retrieve the tasks to use.
This way, the inactive tasks are kept in a multiple-producer/multiple-consumer queue; whenever the queue gets to big, the frontmost (i.e. oldest) connection is closed. Otherwise, each connection is taken, handled and put back at the end of the queue for the next packet. This round-robin should ensure fairness between requests among all clients.